pyproject.toml
[project]
name = "nonebot_plugin_SimpleToWrite"
version = "0.2.0"
description = "为0编程基础的小白提供便捷的功能编写"
authors = [
  { name="STESmly", email="STESmly@mail.com" },
]
readme = "README.md"
requires-python = ">=3.8, <4.0"
dependencies = [
    "nonebot2>=2.2.1",
    "nonebot-adapter-onebot>=2.4.3",
    "matplotlib>=3.7.5",
    "pypinyin>=0.53.0",
    "httpx>=0.27.2",
    "aiohttp>=3.10.10",
    "Brotli>=1.1.0",
    "yarl>=1.15.2",
]

[project.urls]
Homepage = "https://github.com/STESmly/nonebot_plugin_SimpleToWrite"
Repository = "https://github.com/STESmly/nonebot_plugin_SimpleToWrite"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]

[tool.pdm]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
